+++ src/scene.h
@@ -30,7 +30,7 @@ struct Scene {
     std::vector<Mesh> meshes;
     std::vector<Sphere> spheres;
     std::vector<std::variant<PointLight, SegmentLight, ParallelogramLight>> lights;
-    float focalLength = 0;
+    int focalLength = 0;
     float aperture = 0;
     int DOF_samples = 0;
 };
